3 and 4

+++ 0x08-recursion/2-strlen_recursion.c
@@ -7,7 +7,7 @@
  */
 int _strlen_recursion(char *s)
 {
-	long l = 0;
+	int l = 0;
 	if (*s != '\0')
 	{
 		l++;
